Refactored search functionality into multiple Utility classes (task #4333)
